# docs/schema/openapi.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
...
      "ConfigurationCollecte" : {
        "required" : [ "collecteOuverte", "prevenirAnimateurs" ],
        "type" : "object",
        "properties" : {
          "collecteOuverte" : {
            "type" : "boolean"
          },
          "debut" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "fin" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "prevenirAnimateurs" : {
            "type" : "boolean"
          },
          "invitation" : {
            "$ref" : "#/components/schemas/InvitationReport"
          }
        }
      },
      "ConfigurationFoire" : {
        "required" : [ "foireOuverte", "ouverteAujourdhui" ],
        "type" : "object",
        "properties" : {
          "foireOuverte" : {
            "type" : "boolean"
          },
          "debut" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "fin" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "ouverteAujourdhui" : {
            "type" : "boolean"
          }
        }
      },
...
      "Decision" : {
        "type" : "object",
        "properties" : {
          "commentaire" : {
            "type" : "string"
          }
        }
      },
      "Decision1" : {
        "type" : "object",
        "properties" : {
          "commentaire" : {
            "type" : "string"
          }
        }
      },
...
      "InvitationReport" : {
        "type" : "object",
        "required" : [ "envoyes" ],
        "properties" : {
          "envoyes" : {
            "type" : "integer",
            "format" : "int32"
          },
          "sansEmail" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "echecs" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
      "LocalDate" : {
        "type" : "string",
        "format" : "date",
        "examples" : [ "2022-03-10" ]
      },
...
    }
  },
  "paths" : {
...
    "/api/disponibilites/configuration" : {
      "put" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ConfigurationCollecte"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConfigurationCollecte"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Configure",
        "tags" : [ "Declaration Disponibilite Resource" ]
      },
      "get" : {
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConfigurationCollecte"
                }
              }
            }
          }
        },
        "summary" : "Configuration",
        "tags" : [ "Declaration Disponibilite Resource" ]
      }
    },
    "/api/disponibilites/{id}/application" : {
      "post" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "Apply",
        "tags" : [ "Declaration Disponibilite Resource" ]
      }
    },
    "/api/disponibilites/{id}/refus" : {
      "post" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Decision"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Refuse",
        "tags" : [ "Declaration Disponibilite Resource" ]
      }
    },
...
    "/api/echanges/configuration" : {
      "put" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ConfigurationFoire"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConfigurationFoire"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Configure",
        "tags" : [ "Demande Echange Resource" ]
      },
      "get" : {
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConfigurationFoire"
                }
              }
            }
          }
        },
        "summary" : "Configuration",
        "tags" : [ "Demande Echange Resource" ]
      }
    },
    "/api/echanges/{id}/acceptation" : {
      "post" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Decision1"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Accept",
        "tags" : [ "Demande Echange Resource" ]
      }
    },
    "/api/echanges/{id}/impact" : {
      "get" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "Impact",
        "tags" : [ "Demande Echange Resource" ]
      }
    },
    "/api/echanges/{id}/refus" : {
      "post" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Decision1"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Refuse",
        "tags" : [ "Demande Echange Resource" ]
      }
    },
...
  },
  "info" : {
    "title" : "planning-equipes API",
    "version" : "999-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}